requirements.txt
requests
python-dotenv
sendgrid

Flask

# pytest to have the tests run